Added cart button functionality

+++ sephora-backend/Application/Dtos/Rating/CreateRatingDto.cs
@@ -3,7 +3,7 @@ namespace CleanArchitecture.Application.Dtos.Rating;
 public class CreateRatingDto
 {
     public long ProductId { get; set; }
-    public int Rate { get; set; } // 1 to 5
+    public decimal Rate { get; set; } // 1 to 5
     public string? Comment { get; set; }
     
     // The user will be registered in the backend
